+++ forms/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './defaultValidationErrors/defaultValidationErrors.component';
 export * from './defaultValidationErrorsContainer/defaultValidationErrorsContainer.component';
+export * from './reservedSpaceValidationErrorsContainer/reservedSpaceValidationErrorsContainer.component';