+++ forms/src/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './groupHasErrorContainer/groupHasErrorContainer.directive';
+export * from './validationErrorsContainer/validationErrorsContainer.directive';