+++ forms/src/modules/hasError.module.ts
@@ -3,7 +3,8 @@ import {CommonModule} from '@angular/common';
 
 import {HasErrorDirective} from '../directives/hasError/hasError.directive';
 import {GroupHasErrorDirective} from '../directives/groupHasError/groupHasError.directive';
-import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerComponent} from '../components';
+import {GroupHasErrorContainerDirective, ValidationErrorsContainerDirective} from '../directives';
+import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerComponent, ReservedSpaceValidationErrorsContainerComponent} from '../components';
 
 /**
  * Module for input validation directives
@@ -18,13 +19,18 @@ import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerCompon
     [
         HasErrorDirective,
         GroupHasErrorDirective,
+        GroupHasErrorContainerDirective,
+        ValidationErrorsContainerDirective,
         DefaultValidationErrorsComponent,
         DefaultValidationErrorsContainerComponent,
+        ReservedSpaceValidationErrorsContainerComponent,
     ],
     exports: 
     [
         HasErrorDirective,
         GroupHasErrorDirective,
+        GroupHasErrorContainerDirective,
+        ValidationErrorsContainerDirective,
     ]
 })
 export class HasErrorModule